Refactor add-on into modular structure
