+++ __init__.py
@@ -1,3 +1,23 @@
+"""Blender entry point for the DICOMator add-on."""
+from __future__ import annotations
+
+import bpy
+from bpy.props import FloatProperty, PointerProperty
+
+from .constants import MAX_HU_VALUE, MIN_HU_VALUE
+from .dicom_export import export_voxel_grid_to_dicom
+from .operators import MESH_OT_export_dicom
+from .panels import (
+    VIEW3D_PT_dicomator_export_settings,
+    VIEW3D_PT_dicomator_orientation,
+    VIEW3D_PT_dicomator_panel,
+    VIEW3D_PT_dicomator_patient_info,
+    VIEW3D_PT_dicomator_per_object_hu,
+    VIEW3D_PT_dicomator_selection_info,
+)
+from .properties import DICOMatorProperties
+from .voxelization import add_gaussian_noise, voxelize_mesh, voxelize_objects_to_hu
+
 bl_info = {
     "name": "DICOMator",
     "author": "Michael Douglass",
@@ -10,1272 +30,6 @@ bl_info = {
     "category": "3D View",
 }
 
-# Standard library imports
-import os
-import math
-from datetime import datetime
-
-# Blender imports
-import bpy
-from mathutils import Vector
-from bpy.types import Operator, Panel, PropertyGroup
-from bpy.props import (
-    FloatProperty,
-    BoolProperty,
-    IntProperty,
-    PointerProperty,
-    StringProperty,
-    EnumProperty,
-)
-from mathutils.bvhtree import BVHTree
-import numpy as np
-
-# Add these DICOM-specific imports
-try:
-    import pydicom
-    from pydicom.dataset import Dataset, FileDataset
-    from pydicom.uid import generate_uid
-    PYDICOM_AVAILABLE = True
-except ImportError:
-    PYDICOM_AVAILABLE = False
-    print("Warning: pydicom not available. DICOM export functionality will be disabled.")
-
-# Constants
-AIR_DENSITY = -1000.0  # HU value for air (DICOM standard reference for air)
-DEFAULT_DENSITY = 0.0   # Default HU for objects unless overridden per-object
-MAX_HU_VALUE = 3071     # Max HU for 12-bit CT representations
-MIN_HU_VALUE = -1024    # Min HU (typical CT lower bound)
-
-# Safely fetch numeric properties even when Blender returns _PropertyDeferred
-def _get_float_prop(props, name: str, default: float) -> float:
-    # Some Blender properties can be deferred; this wrapper robustly reads them
-    # and always returns a float fallback on failure.
-    try:
-        val = getattr(props, name)
-        return float(val)
-    except Exception:
-        return float(default)
-
-# Safely fetch string properties even when Blender returns _PropertyDeferred
-def _get_str_prop(props, name: str, default: str) -> str:
-    # Robust string getter with safe default for deferred/unset properties.
-    try:
-        val = getattr(props, name)
-        return str(val) if val is not None else str(default)
-    except Exception:
-        return str(default)
-
-def export_voxel_grid_to_dicom(
-        voxel_grid,
-        voxel_size,
-        output_dir,
-        bbox_min,
-        patient_name="Anonymous",
-        patient_id="12345678",
-        patient_sex="M",
-        series_description="CT Series from DICOMator",
-        progress_callback=None,
-        direct_hu=False,
-        patient_position="HFS",
-        # NEW: 4D/temporal and UID controls
-        study_instance_uid=None,
-        frame_of_reference_uid=None,
-        series_instance_uid=None,
-        series_number=1,
-        temporal_position_identifier=None,
-        number_of_temporal_positions=None,
-        phase_index=None,
-        # NEW: old-version temporal tag (add alongside identifier)
-        temporal_position_index=None,
-    ):
-        """
-        Export the voxel grid (binary or HU) as a series of DICOM slices.
-
-        Key points:
-        - If direct_hu=True, we treat the array values as HU directly.
-        - Blender uses meters; DICOM uses millimeters. We convert voxel spacing
-          and origin accordingly.
-        - Rows/Columns: DICOM expects rows=Y, cols=X, so we transpose slices.
-        - Temporal tags: write NumberOfTemporalPositions (total phases),
-          TemporalPositionIndex (timeline frame number), and optionally a
-          TemporalPositionIdentifier (1-based phase index).
-        """
-        if not PYDICOM_AVAILABLE:
-            return {'error': 'pydicom not available'}
-        # Ensure output directory exists
-        os.makedirs(output_dir, exist_ok=True)
-        
-        # Get current date and time
-        current_datetime = datetime.now()
-        date_str = current_datetime.strftime('%Y%m%d')
-        time_str = current_datetime.strftime('%H%M%S.%f')
-        
-        # Dimensions and units
-        num_slices = voxel_grid.shape[2]
-        # Convert to HU grid
-        if direct_hu:
-            hu_grid = np.array(voxel_grid, dtype=np.int16, copy=False)
-        else:
-            hu_grid = np.where(voxel_grid > 0, DEFAULT_DENSITY, AIR_DENSITY).astype(np.int16, copy=False)
-        # Clip values to valid HU range
-        hu_grid = np.clip(hu_grid, MIN_HU_VALUE, MAX_HU_VALUE).astype(np.int16, copy=False)
-
-        # Normalize voxel size to (vx, vy, vz) in meters, then convert to mm
-        if isinstance(voxel_size, (tuple, list, np.ndarray)) and len(voxel_size) == 3:
-            vx_m, vy_m, vz_m = float(voxel_size[0]), float(voxel_size[1]), float(voxel_size[2])
-        else:
-            vx_m = vy_m = vz_m = float(voxel_size)
-        vx_mm, vy_mm, vz_mm = vx_m * 1000.0, vy_m * 1000.0, vz_m * 1000.0
-
-        bbox_min_mm = Vector((bbox_min.x * 1000.0, bbox_min.y * 1000.0, bbox_min.z * 1000.0))
-        
-        # Generate UIDs (reuse if provided)
-        study_instance_uid = study_instance_uid or generate_uid()
-        frame_of_reference_uid = frame_of_reference_uid or generate_uid()
-        series_instance_uid = series_instance_uid or generate_uid()
-        
-        common_metadata = {
-            'study_instance_uid': study_instance_uid,
-            'frame_of_reference_uid': frame_of_reference_uid,
-            'series_instance_uid': series_instance_uid,
-            'series_number': series_number,
-            'series_description': series_description,
-            'patient_name': patient_name,
-            'patient_id': patient_id,
-            'patient_sex': patient_sex,
-            'date_str': date_str,
-            'time_str': time_str,
-        }
-        
-        for i in range(num_slices):
-            # For each axial slice i (Z index), write one DICOM instance.
-            # Note: hu_grid is shaped (width, height, depth). For DICOM, Rows = Y (height), Columns = X (width)
-            slice_data = hu_grid[:, :, i]
-            try:
-                # Create file meta information
-                file_meta = Dataset()
-                file_meta.MediaStorageSOPClassUID = pydicom.uid.CTImageStorage
-                file_meta.MediaStorageSOPInstanceUID = generate_uid()
-                file_meta.ImplementationClassUID = pydicom.uid.PYDICOM_IMPLEMENTATION_UID
-                file_meta.TransferSyntaxUID = pydicom.uid.ExplicitVRLittleEndian
-                
-                # Create the FileDataset instance
-                ds = FileDataset(None, {}, file_meta=file_meta, preamble=b"\0" * 128)
-                
-                # Set DICOM tags
-                ds.ImageType = ['ORIGINAL', 'PRIMARY', 'AXIAL']  # One series per phase (non-multi-frame)
-                ds.SOPClassUID = file_meta.MediaStorageSOPClassUID
-                ds.SOPInstanceUID = file_meta.MediaStorageSOPInstanceUID
-
-                # Patient information
-                ds.PatientName = common_metadata['patient_name']
-                ds.PatientID = common_metadata['patient_id']
-                ds.PatientBirthDate = ''
-                ds.PatientSex = common_metadata['patient_sex']
-                ds.PatientPosition = str(patient_position)
-                
-                # Study information
-                ds.StudyInstanceUID = common_metadata['study_instance_uid']
-                ds.FrameOfReferenceUID = common_metadata['frame_of_reference_uid']
-                ds.StudyID = '1'
-                ds.AccessionNumber = '1'
-                ds.StudyDate = common_metadata['date_str']
-                ds.StudyTime = common_metadata['time_str']
-                ds.ReferringPhysicianName = ''
-                
-                # Series information
-                ds.SeriesInstanceUID = common_metadata['series_instance_uid']
-                ds.SeriesNumber = common_metadata['series_number']
-                ds.SeriesDescription = common_metadata['series_description']
-                ds.SeriesDate = common_metadata['date_str']
-                ds.SeriesTime = common_metadata['time_str']
-                
-                # Temporal (4D) tags (match old version; keep optional identifier for compatibility)
-                # - NumberOfTemporalPositions: total number of exported phases
-                # - TemporalPositionIndex: original timeline frame number
-                # - TemporalPositionIdentifier: 1-based phase index (phase ordering)
-                # Some viewers use one or the other; we include both for compatibility.
-                if number_of_temporal_positions is not None:
-                    ds.NumberOfTemporalPositions = int(number_of_temporal_positions)
-                if temporal_position_index is not None:
-                    ds.TemporalPositionIndex = int(temporal_position_index)
-                if temporal_position_identifier is not None:
-                    ds.TemporalPositionIdentifier = int(temporal_position_identifier)
-                
-                # Equipment information
-                ds.Modality = 'CT'
-                ds.Manufacturer = 'DICOMator'
-                ds.InstitutionName = 'Virtual Hospital'
-                ds.StationName = 'Blender'
-                
-                # Image information
-                ds.InstanceNumber = i + 1
-                # AcquisitionNumber: use phase_index if present, else 1
-                ds.AcquisitionNumber = int(phase_index or 1)
-                ds.ContentDate = common_metadata['date_str']
-                ds.ContentTime = common_metadata['time_str']
-                ds.AcquisitionDate = common_metadata['date_str']
-                ds.AcquisitionTime = common_metadata['time_str']
-                
-                # Image position/orientation with anisotropic Z step
-                ds.ImagePositionPatient = [
-                    float(bbox_min_mm.x),
-                    float(bbox_min_mm.y),
-                    float(bbox_min_mm.z + (i * vz_mm))
-                ]
-                ds.ImageOrientationPatient = [1.0, 0.0, 0.0, 0.0, 1.0, 0.0]
-                ds.SliceLocation = float(bbox_min_mm.z + (i * vz_mm))
-                ds.SliceThickness = float(vz_mm)
-                ds.SpacingBetweenSlices = float(vz_mm)
-                
-                # Pixel data and spacing (rows=Y -> vy_mm, cols=X -> vx_mm)
-                pixel_array = slice_data.T.astype(np.int16, copy=False)
-                rows, cols = pixel_array.shape
-                ds.SamplesPerPixel = 1
-                ds.PhotometricInterpretation = 'MONOCHROME2'
-                ds.Rows = int(rows)
-                ds.Columns = int(cols)
-                ds.PixelSpacing = [float(vy_mm), float(vx_mm)]
-                ds.BitsAllocated = 16
-                ds.BitsStored = 16
-                ds.HighBit = 15
-                ds.PixelRepresentation = 1
-                ds.RescaleIntercept = 0.0
-                ds.RescaleSlope = 1.0
-                ds.WindowCenter = 40
-                ds.WindowWidth = 400
-                ds.PixelData = pixel_array.tobytes()
-
-                # Create filename:
-                # - Prefix with phase index to avoid overwrites across phases.
-                if phase_index is not None:
-                    filename = os.path.join(output_dir, f"Phase_{int(phase_index):03d}_CT_Slice_{i+1:04d}.dcm")
-                else:
-                    filename = os.path.join(output_dir, f"CT_Slice_{i+1:04d}.dcm")
-                
-                # Save the DICOM file
-                ds.is_little_endian = True
-                ds.is_implicit_VR = False
-                ds.save_as(filename)
-                
-                # Progress
-                if progress_callback:
-                    progress_callback(i + 1, num_slices)
-            except Exception as e:
-                print(f"Error saving DICOM file for slice {i+1}: {str(e)}")
-                return {'error': f"Error saving DICOM file for slice {i+1}: {str(e)}"}
-                
-        return {'success': f"Successfully exported {num_slices} DICOM slices to {output_dir}"}
-
-# Helper to build BVH directly from object mesh, optionally through depsgraph (modifiers applied)
-def _bvh_from_object(obj, depsgraph=None, apply_modifiers=False):
-    """
-    Build a BVH in world space.
-    - When apply_modifiers=True and depsgraph is provided, use evaluated mesh (modifiers/shape keys/armature/lattice).
-    - Otherwise, use the object's base mesh (no modifiers).
-    """
-    if apply_modifiers and depsgraph is not None:
-        obj_eval = obj.evaluated_get(depsgraph)
-        me = obj_eval.to_mesh(preserve_all_data_layers=False, depsgraph=depsgraph)
-        try:
-            verts_world = [obj_eval.matrix_world @ v.co for v in me.vertices]
-            polys = [list(p.vertices) for p in me.polygons]
-        finally:
-            obj_eval.to_mesh_clear()
-    else:
-        me = obj.data
-        verts_world = [obj.matrix_world @ v.co for v in me.vertices]
-        polys = [list(p.vertices) for p in me.polygons]
-    return BVHTree.FromPolygons(verts_world, polys)
-
-# Helper to build BVH directly from object mesh (world-space), no depsgraph
-def _bvh_from_object_old(obj):
-    # Build a BVH using the object's base mesh data (object.data) transformed into
-    # world space via matrix_world. This mimics the old code path and does NOT
-    # evaluate modifiers/armatures. For deformed geometry, apply transforms first.
-    me = obj.data
-    verts_world = [obj.matrix_world @ v.co for v in me.vertices]
-    polys = [list(p.vertices) for p in me.polygons]
-    return BVHTree.FromPolygons(verts_world, polys)
-
-def voxelize_mesh(obj, voxel_size=1.0, padding=1):
-    """
-    Voxelize a single mesh by casting vertical (+Z) rays per (X,Y) column.
-    Supports anisotropic voxel sizes: voxel_size = (vx, vy, vz) in meters.
-    """
-    # Normalize voxel size
-    if isinstance(voxel_size, (tuple, list, np.ndarray)) and len(voxel_size) == 3:
-        vx, vy, vz = float(voxel_size[0]), float(voxel_size[1]), float(voxel_size[2])
-    else:
-        vx = vy = vz = float(voxel_size)
-
-    bvh = _bvh_from_object_old(obj)
-
-    # Get object's bounding box in world coordinates
-    bbox_corners = [obj.matrix_world @ Vector(corner) for corner in obj.bound_box]
-    
-    # Calculate bounding box min/max
-    min_x = min(corner.x for corner in bbox_corners)
-    max_x = max(corner.x for corner in bbox_corners)
-    min_y = min(corner.y for corner in bbox_corners)
-    max_y = max(corner.y for corner in bbox_corners)
-    min_z = min(corner.z for corner in bbox_corners)
-    max_z = max(corner.z for corner in bbox_corners)
-    
-    # Add padding per axis
-    min_x -= padding * vx
-    max_x += padding * vx
-    min_y -= padding * vy
-    max_y += padding * vy
-    min_z -= padding * vz
-    max_z += padding * vz
-    
-    # Calculate voxel grid dimensions
-    width = max(1, int(math.ceil((max_x - min_x) / vx)))
-    height = max(1, int(math.ceil((max_y - min_y) / vy)))
-    depth = max(1, int(math.ceil((max_z - min_z) / vz)))
-    
-    # Create voxel array (width, height, depth)
-    voxel_array = np.zeros((width, height, depth), dtype=np.uint8)
-    
-    # Origin point (world coordinates of voxel [0,0,0])
-    origin = Vector((min_x, min_y, min_z))
-    
-    print(f"Voxelizing mesh '{obj.name}' into {width}x{height}x{depth} grid (BVH ray casting)...")
-    
-    # Precompute world coordinates of voxel centers in X and Y
-    xs = min_x + (np.arange(width) + 0.5) * vx
-    ys = min_y + (np.arange(height) + 0.5) * vy
-    
-    # Z center reference for index mapping
-    z0_center = min_z + 0.5 * vz
-    inv_dz = 1.0 / vz
-    ray_dir = Vector((0.0, 0.0, 1.0))
-    max_dist = (max_z - min_z) + 4.0 * vz
-    eps = 1e-6
-    
-    # Iterate over columns (X,Y) and fill Z-intervals using BVH ray casts
-    total_cols = width * height
-    col_count = 0
-    
-    for ix in range(width):
-        xw = float(xs[ix])
-        for iy in range(height):
-            yw = float(ys[iy])
-            # Start ray just below the min Z to ensure first hit is the first surface
-            origin_ray = Vector((xw, yw, min_z - 2.0 * vz))
-            hits_z = []
-            
-            # Collect all intersections along +Z
-            while True:
-                loc, normal, face_index, distance = bvh.ray_cast(origin_ray, ray_dir, max_dist)
-                if loc is None:
-                    break
-                hits_z.append(loc.z)
-                # Move origin slightly above the hit to find the next intersection
-                origin_ray = Vector((loc.x, loc.y, loc.z + eps))
-            
-            if hits_z:
-                hits_z.sort()
-                # Fill every alternating interval [z0, z1)
-                for j in range(0, len(hits_z) - 1, 2):
-                    z0 = hits_z[j]
-                    z1 = hits_z[j + 1]
-                    # Determine voxel indices whose centers lie in (z0, z1)
-                    start_idx = int(math.ceil((z0 - z0_center) * inv_dz))
-                    end_idx = int(math.floor((z1 - z0_center) * inv_dz))
-                    if end_idx >= start_idx:
-                        # Clamp to array bounds
-                        s = max(0, start_idx)
-                        e = min(depth - 1, end_idx)
-                        if e >= s:
-                            voxel_array[ix, iy, s:e + 1] = 1
-            
-            col_count += 1
-            if (col_count % 10000) == 0:
-                print(f"  processed columns: {col_count}/{total_cols} ({(col_count/total_cols)*100:.1f}%)")
-    
-    filled = int(voxel_array.sum())
-    print(f"Voxelization complete. Filled voxels: {filled}/{voxel_array.size}")
-    
-    return voxel_array, origin, (width, height, depth)
-
-# New: voxelize multiple objects into a single HU grid
-def voxelize_objects_to_hu(objects, voxel_size=1.0, padding=1, progress_callback=None, bbox_override=None,
-                           apply_modifiers=False, depsgraph=None):
-    """
-    Voxelize multiple mesh objects into a single HU grid (int16), using per-object HU.
-
-    - If apply_modifiers is True, evaluated meshes are used (modifiers/shape keys/armatures/lattices).
-    - When bbox_override is None, bounds are computed with (or without) modifiers based on apply_modifiers.
-    """
-    # Validate input meshes
-    if not objects:
-        raise ValueError("No objects provided for voxelization")
-
-    if apply_modifiers and depsgraph is None:
-        # Safe fallback
-        depsgraph = bpy.context.evaluated_depsgraph_get()
-
-    # Normalize voxel size (support anisotropic sizes: (vx, vy, vz) in meters)
-    if isinstance(voxel_size, (tuple, list, np.ndarray)) and len(voxel_size) == 3:
-        vx, vy, vz = float(voxel_size[0]), float(voxel_size[1]), float(voxel_size[2])
-    else:
-        vx = vy = vz = float(voxel_size)
-
-    # Compute global bounds across objects (or use provided override).
-    if bbox_override is not None:
-        min_x, max_x, min_y, max_y, min_z, max_z = bbox_override
-    else:
-        # Compute bounds in world space (evaluated if requested)
-        min_x = min_y = min_z = float('inf')
-        max_x = max_y = max_z = float('-inf')
-        for obj in objects:
-            if apply_modifiers and depsgraph is not None:
-                obj_eval = obj.evaluated_get(depsgraph)
-                me = obj_eval.to_mesh(preserve_all_data_layers=False, depsgraph=depsgraph)
-                try:
-                    for v in me.vertices:
-                        vw = obj_eval.matrix_world @ v.co
-                        if vw.x < min_x: min_x = vw.x
-                        if vw.x > max_x: max_x = vw.x
-                        if vw.y < min_y: min_y = vw.y
-                        if vw.y > max_y: max_y = vw.y
-                        if vw.z < min_z: min_z = vw.z
-                        if vw.z > max_z: max_z = vw.z
-                finally:
-                    obj_eval.to_mesh_clear()
-            else:
-                for corner in obj.bound_box:
-                    cw = obj.matrix_world @ Vector(corner)
-                    if cw.x < min_x: min_x = cw.x
-                    if cw.x > max_x: max_x = cw.x
-                    if cw.y < min_y: min_y = cw.y
-                    if cw.y > max_y: max_y = cw.y
-                    if cw.z < min_z: min_z = cw.z
-                    if cw.z > max_z: max_z = cw.z
-        # Add padding per axis only when not overriding
-        min_x -= padding * vx
-        max_x += padding * vx
-        min_y -= padding * vy
-        max_y += padding * vy
-        min_z -= padding * vz
-        max_z += padding * vz
-
-    # Grid dimensions and origin
-    width = max(1, int(math.ceil((max_x - min_x) / vx)))
-    height = max(1, int(math.ceil((max_y - min_y) / vy)))
-    depth = max(1, int(math.ceil((max_z - min_z) / vz)))
-    origin = Vector((min_x, min_y, min_z))
-
-    print(f"Voxelizing {len(objects)} objects into {width}x{height}x{depth} HU grid...")
-
-    hu_array = np.full((width, height, depth), int(AIR_DENSITY), dtype=np.int16)
-
-    # Precompute voxel centers
-    xs = min_x + (np.arange(width) + 0.5) * vx
-    ys = min_y + (np.arange(height) + 0.5) * vy
-    z0_center = min_z + 0.5 * vz
-    inv_dz = 1.0 / vz
-    ray_dir = Vector((0.0, 0.0, 1.0))
-    max_dist = (max_z - min_z) + 4.0 * vz
-    eps = 1e-6
-
-    # Prepare per-object (BVH, HU, name) tuples with alphabetical priority.
-    # Sort by name (case-insensitive). Later entries have higher priority (overwrite on overlap).
-    obj_data = []
-    sorted_objs = sorted(objects, key=lambda o: o.name.casefold())
-    for obj in sorted_objs:
-        bvh = _bvh_from_object(obj, depsgraph=depsgraph, apply_modifiers=apply_modifiers)
-        hu_val = float(getattr(obj, 'dicomator_hu', DEFAULT_DENSITY))
-        hu_val = max(MIN_HU_VALUE, min(MAX_HU_VALUE, hu_val))
-        obj_data.append((bvh, np.int16(hu_val), obj.name))
-
-    total_cols = width * height
-    col_count = 0
-
-    # Iterate columns and fill Z-intervals per object, taking maximum HU on overlaps.
-    # Progress callback can be hooked into Blender's progress bar.
-    for ix in range(width):
-        xw = float(xs[ix])
-        for iy in range(height):
-            yw = float(ys[iy])
-            for bvh, hu_val, _name in obj_data:
-                origin_ray = Vector((xw, yw, min_z - 2.0 * vz))
-                hits_z = []
-                while True:
-                    loc, normal, face_index, distance = bvh.ray_cast(origin_ray, ray_dir, max_dist)
-                    if loc is None:
-                        break
-                    hits_z.append(loc.z)
-                    origin_ray = Vector((loc.x, loc.y, loc.z + eps))
-                if hits_z:
-                    hits_z.sort()
-                    for j in range(0, len(hits_z) - 1, 2):
-                        z0 = hits_z[j]
-                        z1 = hits_z[j + 1]
-                        start_idx = int(math.ceil((z0 - z0_center) * inv_dz))
-                        end_idx = int(math.floor((z1 - z0_center) * inv_dz))
-                        if end_idx >= start_idx:
-                            s = max(0, start_idx)
-                            e = min(depth - 1, end_idx)
-                            if e >= s:
-                                # Alphabetical overwrite priority: assign directly so later names win.
-                                hu_array[ix, iy, s:e+1] = hu_val
-            col_count += 1
-            if progress_callback and (col_count % 5000) == 0:
-                progress_callback(col_count, total_cols)
-            if (col_count % 20000) == 0:
-                print(f"  processed columns: {col_count}/{total_cols} ({(col_count/total_cols)*100:.1f}%)")
-
-    print("Voxelization complete (multi-object HU grid).")
-    return hu_array, origin, (width, height, depth)
-
-# Helper: add Gaussian noise (in HU) to a HU grid
-def add_gaussian_noise(hu_array: np.ndarray, std_hu: float) -> np.ndarray:
-    """
-    Add zero-mean Gaussian noise with standard deviation std_hu (HU) to the HU grid.
-    The result is clipped to the valid HU range and returned as int16.
-    """
-    if std_hu <= 0.0:
-        return hu_array.astype(np.int16, copy=False)
-    noisy = hu_array.astype(np.float32, copy=False) + np.random.normal(0.0, float(std_hu), hu_array.shape).astype(np.float32, copy=False)
-    noisy = np.clip(noisy, MIN_HU_VALUE, MAX_HU_VALUE)
-    return noisy.astype(np.int16, copy=False)
-
-class MESH_OT_export_dicom(Operator):
-    """Export mesh to DICOM files"""
-    bl_idname = "mesh.export_dicom"
-    bl_label = "Export to DICOM"
-    bl_options = {'REGISTER', 'UNDO'}
-    
-    @classmethod
-    def poll(cls, context):
-        return (context.active_object is not None and 
-                context.active_object.type == 'MESH' and
-                context.active_object.mode == 'OBJECT')
-    
-    def execute(self, context):
-        # Guardrails: require pydicom and at least one mesh selected.
-        if not PYDICOM_AVAILABLE:
-            self.report({'ERROR'}, "pydicom library not available")
-            return {'CANCELLED'}
-        
-        # Collect all selected mesh objects
-        selected_meshes = [o for o in context.selected_objects if o.type == 'MESH']
-        if not selected_meshes:
-            self.report({'ERROR'}, "Please select at least one mesh object")
-            return {'CANCELLED'}
-        
-        # Use active object for mode validation but export all selected meshes
-        obj = context.active_object
-        if not obj or obj.type != 'MESH':
-            self.report({'ERROR'}, "Please select a mesh object")
-            return {'CANCELLED'}
-        
-        # Get properties from the panel
-        props = context.scene.dicomator_props
-        
-        # Check if export directory is specified
-        output_dir = props.export_directory
-        if not output_dir:
-            self.report({'ERROR'}, "Please specify an export directory")
-            return {'CANCELLED'}
-        
-        # Convert Blender relative paths to absolute paths
-        if output_dir.startswith('//'):
-            # Blender relative path - convert to absolute
-            relative_path = output_dir[2:]  # Remove '//' prefix
-            
-            # Normalize path separators for the current OS
-            relative_path = relative_path.replace('/', os.sep).replace('\\', os.sep)
-            
-            if bpy.data.filepath:
-                # If blend file is saved, make path relative to blend file
-                blend_dir = os.path.dirname(bpy.data.filepath)
-                output_dir = os.path.join(blend_dir, relative_path)
-            else:
-                # If blend file is not saved, use current working directory
-                output_dir = os.path.join(os.getcwd(), relative_path)
-        
-        # Ensure the path is absolute and normalize it
-        output_dir = os.path.abspath(os.path.normpath(output_dir))
-        
-        # Create directory if it doesn't exist
-        try:
-            os.makedirs(output_dir, exist_ok=True)
-            self.report({'INFO'}, f"Using export directory: {output_dir}")
-        except Exception as e:
-            self.report({'ERROR'}, f"Cannot create output directory: {str(e)}")
-            return {'CANCELLED'}
-        
-        # Verify directory is writable
-        if not os.access(output_dir, os.W_OK):
-            self.report({'ERROR'}, f"Output directory is not writable: {output_dir}")
-            return {'CANCELLED'}
-        
-        try:
-            # Effective resolutions (mm) with fallback to legacy grid_resolution
-            props = context.scene.dicomator_props
-            lateral_mm = _get_float_prop(props, "lateral_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-            axial_mm = _get_float_prop(props, "axial_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-            vx_m = float(lateral_mm) * 0.001
-            vy_m = float(lateral_mm) * 0.001
-            vz_m = float(axial_mm) * 0.001
-
-            # Collect 4D frame parameters
-            props = context.scene.dicomator_props
-            export_4d = bool(props.export_4d)
-            apply_modifiers = bool(getattr(props, "apply_modifiers", True))
-            if export_4d:
-                if props.use_timeline_range:
-                    start = int(context.scene.frame_start)
-                    end = int(context.scene.frame_end)
-                else:
-                    start = int(props.frame_start)
-                    end = int(props.frame_end)
-                step = max(1, int(props.frame_step))
-                if end < start:
-                    self.report({'ERROR'}, "End frame must be >= start frame")
-                    return {'CANCELLED'}
-
-            # Compute bounds (global across frames if 4D), respecting modifiers when requested
-            padding = 1
-            saved_frame = context.scene.frame_current
-            try:
-                if export_4d:
-                    min_x = min_y = min_z = float('inf')
-                    max_x = max_y = max_z = float('-inf')
-                    for f in range(start, end + 1, step):
-                        context.scene.frame_set(f, subframe=0.0)
-                        context.scene.frame_current = f
-                        _force_ui_redraw()
-                        depsgraph = context.evaluated_depsgraph_get()
-                        print(f"[DICOMator] Bounds pass at frame {f}")
-                        for o in selected_meshes:
-                            if apply_modifiers:
-                                o_eval = o.evaluated_get(depsgraph)
-                                me = o_eval.to_mesh(preserve_all_data_layers=False, depsgraph=depsgraph)
-                                try:
-                                    for v in me.vertices:
-                                        vw = o_eval.matrix_world @ v.co
-                                        if vw.x < min_x: min_x = vw.x
-                                        if vw.x > max_x: max_x = vw.x
-                                        if vw.y < min_y: min_y = vw.y
-                                        if vw.y > max_y: max_y = vw.y
-                                        if vw.z < min_z: min_z = vw.z
-                                        if vw.z > max_z: max_z = vw.z
-                                finally:
-                                    o_eval.to_mesh_clear()
-                            else:
-                                for corner in o.bound_box:
-                                    cw = o.matrix_world @ Vector(corner)
-                                    if cw.x < min_x: min_x = cw.x
-                                    if cw.x > max_x: max_x = cw.x
-                                    if cw.y < min_y: min_y = cw.y
-                                    if cw.y > max_y: max_y = cw.y
-                                    if cw.z < min_z: min_z = cw.z
-                                    if cw.z > max_z: max_z = cw.z
-                else:
-                    # Single-frame bounds
-                    depsgraph = context.evaluated_depsgraph_get()
-                    min_x = min_y = min_z = float('inf')
-                    max_x = max_y = max_z = float('-inf')
-                    for o in selected_meshes:
-                        if apply_modifiers:
-                            o_eval = o.evaluated_get(depsgraph)
-                            me = o_eval.to_mesh(preserve_all_data_layers=False, depsgraph=depsgraph)
-                            try:
-                                for v in me.vertices:
-                                    vw = o_eval.matrix_world @ v.co
-                                    if vw.x < min_x: min_x = vw.x
-                                    if vw.x > max_x: max_x = vw.x
-                                    if vw.y < min_y: min_y = vw.y
-                                    if vw.y > max_y: max_y = vw.y
-                                    if vw.z < min_z: min_z = vw.z
-                                    if vw.z > max_z: max_z = vw.z
-                            finally:
-                                o_eval.to_mesh_clear()
-                        else:
-                            for corner in o.bound_box:
-                                cw = o.matrix_world @ Vector(corner)
-                                if cw.x < min_x: min_x = cw.x
-                                if cw.x > max_x: max_x = cw.x
-                                if cw.y < min_y: min_y = cw.y
-                                if cw.y > max_y: max_y = cw.y
-                                if cw.z < min_z: min_z = cw.z
-                                if cw.z > max_z: max_z = cw.z
-            finally:
-                context.scene.frame_set(saved_frame)
-
-            # Calculate selection dimensions
-            obj_width = max_x - min_x
-            obj_height = max_y - min_y
-            obj_depth = max_z - min_z
-
-            # Estimate voxel grid size (with anisotropic padding)
-            estimated_width = int(math.ceil((obj_width + 2 * padding * vx_m) / vx_m))
-            estimated_height = int(math.ceil((obj_height + 2 * padding * vy_m) / vy_m))
-            estimated_depth = int(math.ceil((obj_depth + 2 * padding * vz_m) / vz_m))
-
-            # Safety check (non-blocking: warn but continue)
-            max_voxels_per_dimension = 2000
-            total_estimated_voxels = estimated_width * estimated_height * estimated_depth
-            max_total_voxels = 100_000_000  # ~200MB int16
-
-            if (estimated_width > max_voxels_per_dimension or 
-                estimated_height > max_voxels_per_dimension or 
-                estimated_depth > max_voxels_per_dimension or
-                total_estimated_voxels > max_total_voxels):
-                self.report({'WARNING'}, 
-                    f"Voxel grid very large: {estimated_width}x{estimated_height}x{estimated_depth} "
-                    f"({total_estimated_voxels:,} voxels). "
-                    f"Selection size: {obj_width:.3f}x{obj_height:.3f}x{obj_depth:.3f}m. "
-                    f"Continuing anyway. This may be slow or run out of memory. "
-                    f"Consider increasing resolution (current: lateral {lateral_mm}mm, axial {axial_mm}mm).")
-
-            # Progress callback function
-            def progress_callback(current, total):
-                progress = min(1.0, max(0.0, current / max(1, total)))
-                context.window_manager.progress_update(progress)
-
-            # Start progress indicator
-            context.window_manager.progress_begin(0, 1)
-
-            if not export_4d:
-                # Single-frame export:
-                # - Voxelize all selected meshes into one HU grid.
-                # - Export as a single DICOM series with current patient/series info.
-                self.report({'INFO'}, 
-                    f"Voxelizing {len(selected_meshes)} mesh(es) with lateral {lateral_mm}mm, axial {axial_mm}mm. "
-                    f"Estimated grid: {estimated_width}x{estimated_height}x{estimated_depth}")
-
-                hu_array, origin, dimensions = voxelize_objects_to_hu(
-                    selected_meshes,
-                    voxel_size=(vx_m, vy_m, vz_m),
-                    padding=padding,
-                    progress_callback=progress_callback,
-                    apply_modifiers=apply_modifiers,
-                    depsgraph=context.evaluated_depsgraph_get(),
-                )
-
-                # Optionally add Gaussian noise in HU before export
-                if getattr(props, "enable_noise", False) and _get_float_prop(props, "noise_std_dev_hu", 0.0) > 0.0:
-                    hu_array_to_export = add_gaussian_noise(hu_array, _get_float_prop(props, "noise_std_dev_hu", 0.0))
-                else:
-                    hu_array_to_export = hu_array
-
-                result = export_voxel_grid_to_dicom(
-                    hu_array_to_export,
-                    (vx_m, vy_m, vz_m),
-                    output_dir,
-                    origin,
-                    patient_name=props.patient_name,
-                    patient_id=props.patient_id,
-                    patient_sex=props.patient_sex,
-                    series_description=props.series_description,
-                    progress_callback=progress_callback,
-                    direct_hu=True,
-                    patient_position=props.patient_position,
-                )
-                context.window_manager.progress_end()
-                if 'error' in result:
-                    self.report({'ERROR'}, result['error'])
-                    return {'CANCELLED'}
-                else:
-                    self.report({'INFO'}, result['success'])
-                    return {'FINISHED'}
-
-            # 4D export path - explicit timeline loop (old methodology)
-            frames = list(range(start, end + 1, step))
-            num_phases = len(frames)
-            self.report({'INFO'}, 
-                f"Exporting {num_phases} phase(s) with lateral {lateral_mm}mm, axial {axial_mm}mm. "
-                f"Grid: {estimated_width}x{estimated_height}x{estimated_depth}")
-
-            # Stable bbox across phases (padding applied once)
-            min_x_p = min_x - padding * vx_m
-            max_x_p = max_x + padding * vx_m
-            min_y_p = min_y - padding * vy_m
-            max_y_p = max_y + padding * vy_m
-            min_z_p = min_z - padding * vz_m
-            max_z_p = max_z + padding * vz_m
-            bbox_override = (min_x_p, max_x_p, min_y_p, max_y_p, min_z_p, max_z_p)
-
-            study_uid = generate_uid()
-            frame_of_ref_uid = generate_uid()
-
-            saved_frame = context.scene.frame_current
-            try:
-                for phase_index, frame in enumerate(frames, start=1):
-                    # Move timeline to the correct frame and force redraw for visible feedback.
-                    context.scene.frame_set(frame, subframe=0.0)
-                    context.scene.frame_current = frame
-                    _force_ui_redraw()
-
-                    # Build HU grid for this phase at fixed bbox; no depsgraph usage.
-                    hu_array, origin, dimensions = voxelize_objects_to_hu(
-                        selected_meshes,
-                        voxel_size=(vx_m, vy_m, vz_m),
-                        padding=0,
-                        progress_callback=progress_callback,
-                        bbox_override=bbox_override,
-                        apply_modifiers=apply_modifiers,
-                        depsgraph=context.evaluated_depsgraph_get(),
-                    )
-
-                    # Optionally add Gaussian noise per phase
-                    if getattr(props, "enable_noise", False) and _get_float_prop(props, "noise_std_dev_hu", 0.0) > 0.0:
-                        hu_array_to_export = add_gaussian_noise(hu_array, _get_float_prop(props, "noise_std_dev_hu", 0.0))
-                    else:
-                        hu_array_to_export = hu_array
-
-                    percent = (phase_index / num_phases) * 100.0
-                    phase_series_desc = f"{props.series_description} - Phase {phase_index} ({percent:.1f}%)"
-                    series_uid = generate_uid()
-                    series_number = phase_index
-
-                    result = export_voxel_grid_to_dicom(
-                        hu_array_to_export,
-                        (vx_m, vy_m, vz_m),
-                        output_dir,
-                        origin,
-                        patient_name=props.patient_name,
-                        patient_id=props.patient_id,
-                        patient_sex=props.patient_sex,
-                        series_description=phase_series_desc,
-                        progress_callback=progress_callback,
-                        direct_hu=True,
-                        patient_position=props.patient_position,
-                        study_instance_uid=study_uid,
-                        frame_of_reference_uid=frame_of_ref_uid,
-                        series_instance_uid=series_uid,
-                        series_number=series_number,
-                        number_of_temporal_positions=num_phases,
-                        temporal_position_index=frame,
-                        temporal_position_identifier=phase_index,
-                        phase_index=phase_index,
-                    )
-                    if 'error' in result:
-                        context.window_manager.progress_end()
-                        self.report({'ERROR'}, f"Phase {phase_index}: {result['error']}")
-                        return {'CANCELLED'}
-            finally:
-                # Restore the user's original frame after export completes or on error.
-                context.scene.frame_set(saved_frame)
-
-            context.window_manager.progress_end()
-            self.report({'INFO'}, f"Successfully exported {num_phases} phase(s) to {output_dir}")
-            return {'FINISHED'}
-
-        except Exception as e:
-            # Ensure progress is ended and show the error to the user.
-            context.window_manager.progress_end()
-            self.report({'ERROR'}, f"Export failed: {str(e)}")
-            return {'CANCELLED'}
-
-
-class DICOMatorProperties(PropertyGroup):
-    """Properties for DICOMator add-on"""
-    # Switch to annotation style so Blender 4.x registers RNA correctly
-    patient_name: bpy.props.StringProperty(
-        name="Patient Name",
-        description="Name of the patient",
-        default="Anonymous"
-    )
-    patient_id: bpy.props.StringProperty(
-        name="MRN",
-        description="Medical Record Number (Patient ID)",
-        default="12345678"
-    )
-    patient_sex: bpy.props.EnumProperty(
-        name="Patient Sex",
-        description="Patient sex",
-        items=[
-            ('M', 'Male', 'Male'),
-            ('F', 'Female', 'Female'),
-            ('O', 'Other', 'Other'),
-        ],
-        default='M'
-    )
-    # NEW: DICOM PatientPosition (orientation)
-    patient_position: bpy.props.EnumProperty(
-        name="Patient Position",
-        description="DICOM Patient Position (image orientation context)",
-        items=[
-            ('HFS', 'Head First Supine', 'Head First Supine'),
-            ('FFS', 'Feet First Supine', 'Feet First Supine'),
-            ('HFP', 'Head First Prone', 'Head First Prone'),
-            ('FFP', 'Feet First Prone', 'Feet First Prone'),
-            ('HFDR', 'Head First Decubitus Right', 'Head First Decubitus Right'),
-            ('HFDL', 'Head First Decubitus Left', 'Head First Decubitus Left'),
-            ('FFDR', 'Feet First Decubitus Right', 'Feet First Decubitus Right'),
-            ('FFDL', 'Feet First Decubitus Left', 'Feet First Decubitus Left'),
-        ],
-        default='HFS'
-    )
-    # NEW: 4D export controls
-    export_4d: bpy.props.BoolProperty(
-        name="Export 4D (use animated frames)",
-        description="Export each selected animation frame as a 4D CT phase",
-        default=False
-    )
-    use_timeline_range: bpy.props.BoolProperty(
-        name="Use Timeline Range",
-        description="Use the scene's timeline start/end as the frame range",
-        default=True
-    )
-    frame_start: bpy.props.IntProperty(
-        name="Start Frame",
-        description="First frame to export (when not using timeline range)",
-        default=1,
-        min=1
-    )
-    frame_end: bpy.props.IntProperty(
-        name="End Frame",
-        description="Last frame to export (when not using timeline range)",
-        default=250,
-        min=1
-    )
-    frame_step: bpy.props.IntProperty(
-        name="Frame Step",
-        description="Step between frames to export",
-        default=1,
-        min=1
-    )
-    lateral_resolution_mm: bpy.props.FloatProperty(
-        name="Lateral Resolution (mm)",
-        description="Voxel size in X/Y (mm)",
-        default=2.0,
-        min=0.1,
-        max=10.0,
-        step=10,
-        precision=2
-    )
-    axial_resolution_mm: bpy.props.FloatProperty(
-        name="Axial Resolution (mm)",
-        description="Voxel size in Z (mm)",
-        default=2.0,
-        min=0.1,
-        max=10.0,
-        step=10,
-        precision=2
-    )
-    # Apply modifiers/deformations toggle
-    apply_modifiers: bpy.props.BoolProperty(
-        name="Apply Modifiers/Deformations",
-        description="Evaluate modifiers/shape keys/armatures/lattices when voxelizing",
-        default=True
-    )
-    export_directory: bpy.props.StringProperty(
-        name="Export Directory",
-        description="Directory to save DICOM files",
-        subtype='DIR_PATH',
-        default="C:\\Users\\Public\\DICOM_Export"
-    )
-    series_description: bpy.props.StringProperty(
-        name="Series Description",
-        description="Description for the DICOM series",
-        default="CT Series from DICOMator"
-    )
-    # Gaussian noise controls (in HU)
-    enable_noise: bpy.props.BoolProperty(
-        name="Add Gaussian Noise",
-        description="Add zero-mean Gaussian noise to exported HU images",
-        default=False
-    )
-    noise_std_dev_hu: bpy.props.FloatProperty(
-        name="Noise Std. Dev. (HU)",
-        description="Standard deviation of Gaussian noise in Hounsfield Units",
-        default=20.0,
-        min=0.0,
-        soft_max=500.0,
-        step=10,
-        precision=1
-    )
-
-
-class VIEW3D_PT_dicomator_panel(Panel):
-    """DICOMator panel in the 3D viewport sidebar"""
-    bl_label = "DICOMator"
-    bl_idname = "VIEW3D_PT_dicomator_panel"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    
-    def draw(self, context):
-        layout = self.layout
-        # Keep parent minimal; show message if nothing selected
-        if not (context.active_object and context.active_object.type == 'MESH'):
-            layout.label(text="Select a mesh object to export", icon='INFO')
-        else:
-            layout.label(text="Expand sections below to configure export", icon='TRIA_DOWN')
-
-
-# NEW: Collapsible subpanels
-class VIEW3D_PT_dicomator_selection_info(Panel):
-    bl_label = "Selection Info"
-    bl_idname = "VIEW3D_PT_dicomator_selection_info"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    bl_parent_id = "VIEW3D_PT_dicomator_panel"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        props = context.scene.dicomator_props
-        if not (context.active_object and context.active_object.type == 'MESH'):
-            layout.label(text="No mesh selected", icon='INFO')
-            return
-
-        # Selected meshes and active
-        selected_meshes = [o for o in context.selected_objects if o.type == 'MESH']
-        obj = context.active_object
-        sel_count = len(selected_meshes)
-        if sel_count > 1:
-            layout.label(text=f"Selected: {sel_count} meshes (Active: {obj.name})", icon='MESH_DATA')
-        else:
-            layout.label(text=f"Selected: {obj.name}", icon='MESH_DATA')
-
-        # Combined bounds
-        bbox_corners = []
-        for o in selected_meshes:
-            bbox_corners.extend([o.matrix_world @ Vector(corner) for corner in o.bound_box])
-        min_x = min(corner.x for corner in bbox_corners)
-        max_x = max(corner.x for corner in bbox_corners)
-        min_y = min(corner.y for corner in bbox_corners)
-        max_y = max(corner.y for corner in bbox_corners)
-        min_z = min(corner.z for corner in bbox_corners)
-        max_z = max(corner.z for corner in bbox_corners)
-
-        obj_width = max_x - min_x
-        obj_height = max_y - min_y
-        obj_depth = max_z - min_z
-
-        box = layout.box()
-        box.label(text="Selection Info", icon='INFO')
-        box.label(text=f"Size: {obj_width:.2f} x {obj_height:.2f} x {obj_depth:.2f} m")
-
-        # Voxel grid estimate (anisotropic)
-        props = context.scene.dicomator_props
-        lateral_mm = _get_float_prop(props, "lateral_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-        axial_mm = _get_float_prop(props, "axial_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-        if lateral_mm > 0.0 and axial_mm > 0.0:
-            vx = lateral_mm * 0.001
-            vy = lateral_mm * 0.001
-            vz = axial_mm * 0.001
-            padding = 1
-            est_width = int(math.ceil((obj_width + 2 * padding * vx) / vx))
-            est_height = int(math.ceil((obj_height + 2 * padding * vy) / vy))
-            est_depth = int(math.ceil((obj_depth + 2 * padding * vz) / vz))
-            total_voxels = est_width * est_height * est_depth
-
-            box.label(text=f"Est. Grid: {est_width} x {est_height} x {est_depth}")
-            box.label(text=f"Total Voxels: {total_voxels:,}")
-
-            memory_mb = (total_voxels * 2) / (1024 * 1024)
-            box.label(text=f"Est. Memory: {memory_mb:.1f} MB")
-
-            if total_voxels > 100_000_000:
-                box.label(text="Grid too large!", icon='CANCEL')
-            elif total_voxels > 50_000_000:
-                box.label(text="Large grid - may be slow", icon='ERROR')
-
-
-class VIEW3D_PT_dicomator_per_object_hu(Panel):
-    bl_label = "Per-Object HU"
-    bl_idname = "VIEW3D_PT_dicomator_per_object_hu"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    bl_parent_id = "VIEW3D_PT_dicomator_panel"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        if not (context.active_object and context.active_object.type == 'MESH'):
-            layout.label(text="No mesh selected", icon='INFO')
-            return
-        selected_meshes = [o for o in context.selected_objects if o.type == 'MESH']
-        box_hu = layout.box()
-        box_hu.label(text="Per-Object HU", icon='MOD_PHYSICS')
-        if selected_meshes:
-            for o in selected_meshes:
-                row = box_hu.row(align=True)
-                row.prop(o, "dicomator_hu", text=f"{o.name} HU")
-
-
-class VIEW3D_PT_dicomator_patient_info(Panel):
-    bl_label = "Patient Information"
-    bl_idname = "VIEW3D_PT_dicomator_patient_info"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    bl_parent_id = "VIEW3D_PT_dicomator_panel"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        props = context.scene.dicomator_props
-        box = layout.box()
-        box.label(text="Patient Information", icon='USER')
-        box.prop(props, "patient_name")
-        box.prop(props, "patient_id")
-        box.prop(props, "patient_sex")
-
-
-class VIEW3D_PT_dicomator_orientation(Panel):
-    bl_label = "Image Orientation"
-    bl_idname = "VIEW3D_PT_dicomator_orientation"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    bl_parent_id = "VIEW3D_PT_dicomator_panel"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        props = context.scene.dicomator_props
-        box_or = layout.box()
-        box_or.label(text="Image Orientation", icon='ORIENTATION_GIMBAL')
-        box_or.prop(props, "patient_position", text="Patient Position")
-
-
-class VIEW3D_PT_dicomator_export_settings(Panel):
-    bl_label = "Export Settings"
-    bl_idname = "VIEW3D_PT_dicomator_export_settings"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = "DICOMator"
-    bl_parent_id = "VIEW3D_PT_dicomator_panel"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    def draw(self, context):
-        layout = self.layout
-        props = context.scene.dicomator_props
-
-        box = layout.box()
-        box.label(text="Export Settings", icon='SETTINGS')
-
-        # NEW: show anisotropic resolution
-        row = box.row(align=True)
-        row.prop(props, "lateral_resolution_mm", text="Lateral (mm)")
-        row.prop(props, "axial_resolution_mm", text="Axial (mm)")
-
-        box.prop(props, "apply_modifiers", text="Apply Modifiers/Deformations")
-        box.prop(props, "export_directory")
-
-        # NEW: 4D controls
-        col = box.column(align=True)
-        col.prop(props, "export_4d")
-        if props.export_4d:
-            row = col.row(align=True)
-            row.prop(props, "use_timeline_range")
-            if props.use_timeline_range:
-                row = col.row(align=True)
-                row.label(text=f"Timeline: {context.scene.frame_start} → {context.scene.frame_end}", icon='TIME')
-            else:
-                row = col.row(align=True)
-                row.prop(props, "frame_start")
-                row.prop(props, "frame_end")
-            col.prop(props, "frame_step")
-
-        # Resolved export path and description
-        export_dir_val = _get_str_prop(props, "export_directory", "")
-        if export_dir_val.startswith('//'):
-            relative_path = export_dir_val[2:].replace('/', os.sep).replace('\\', os.sep)
-            if bpy.data.filepath:
-                blend_dir = os.path.dirname(bpy.data.filepath)
-                resolved_path = os.path.join(blend_dir, relative_path)
-            else:
-                resolved_path = os.path.join(os.getcwd(), relative_path)
-            resolved_path = os.path.abspath(os.path.normpath(resolved_path))
-            box.label(text=f"Resolved: {resolved_path}", icon='FILE_FOLDER')
-
-        box.prop(props, "series_description")
-
-        # Gaussian noise UI
-        noise_box = box.box()
-        noise_box.label(text="Gaussian Noise", icon='RNDCURVE')
-        row = noise_box.row()
-        row.prop(props, "enable_noise")
-        if props.enable_noise:
-            row = noise_box.row(align=True)
-            row.prop(props, "noise_std_dev_hu", text="Std. Dev. (HU)")
-
-        # Export button and guardrails
-        if PYDICOM_AVAILABLE:
-            export_dir = _get_str_prop(props, "export_directory", "")
-            if export_dir.startswith('//'):
-                relative_path = export_dir[2:].replace('/', os.sep).replace('\\', os.sep)
-                if bpy.data.filepath:
-                    blend_dir = os.path.dirname(bpy.data.filepath)
-                    export_dir = os.path.join(blend_dir, relative_path)
-                else:
-                    export_dir = os.path.join(os.getcwd(), relative_path)
-                export_dir = os.path.abspath(os.path.normpath(export_dir))
-
-            if export_dir and export_dir.strip():
-                if context.active_object and context.active_object.type == 'MESH':
-                    selected_meshes = [o for o in context.selected_objects if o.type == 'MESH']
-                    # For gating, keep single-frame estimate as before (4D sizing is handled at export time)
-                    bbox_corners = []
-                    for o in selected_meshes:
-                        bbox_corners.extend([o.matrix_world @ Vector(corner) for corner in o.bound_box])
-                    min_x = min(corner.x for corner in bbox_corners)
-                    max_x = max(corner.x for corner in bbox_corners)
-                    min_y = min(corner.y for corner in bbox_corners)
-                    max_y = max(corner.y for corner in bbox_corners)
-                    min_z = min(corner.z for corner in bbox_corners)
-                    max_z = max(corner.z for corner in bbox_corners)
-                    obj_width = max_x - min_x
-                    obj_height = max_y - min_y
-                    obj_depth = max_z - min_z
-
-                    lateral_mm2 = _get_float_prop(props, "lateral_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-                    axial_mm2 = _get_float_prop(props, "axial_resolution_mm", _get_float_prop(props, "grid_resolution", 2.0))
-                    if lateral_mm2 > 0.0 and axial_mm2 > 0.0:
-                        vx = lateral_mm2 * 0.001
-                        vy = lateral_mm2 * 0.001
-                        vz = axial_mm2 * 0.001
-                        est_total = int(math.ceil((obj_width + 2 * vx) / vx)) * \
-                                    int(math.ceil((obj_height + 2 * vy) / vy)) * \
-                                    int(math.ceil((obj_depth + 2 * vz) / vz))
-
-                        if est_total > 100_000_000:
-                            layout.label(text="Grid very large - may be slow or fail (memory)", icon='ERROR')
-                        # Always allow export
-                        layout.operator("mesh.export_dicom", text="Export to DICOM", icon='EXPORT')
-                    else:
-                        layout.operator("mesh.export_dicom", text="Export to DICOM", icon='EXPORT')
-                else:
-                    layout.label(text="Select a mesh object to export", icon='INFO')
-            else:
-                layout.label(text="Please select export directory", icon='INFO')
-        else:
-            # ...existing pydicom missing UI...
-            pass
-
-
-# Helper: force UI redraw so the timeline visibly advances during 4D export
-def _force_ui_redraw():
-    # Some long-running operators block UI refresh. This function nudges
-    # Blender to update the timeline and viewport so users can see progress.
-    try:
-        bpy.ops.wm.redraw_timer(type='DRAW_WIN', iterations=1)
-    except Exception:
-        wm = bpy.context.window_manager
-        if wm:
-            for window in wm.windows:
-                for area in window.screen.areas:
-                    area.tag_redraw()
-
-
-
-
-# --- Registration ---
-
 classes = (
     DICOMatorProperties,
     MESH_OT_export_dicom,
@@ -1287,11 +41,11 @@ classes = (
     VIEW3D_PT_dicomator_export_settings,
 )
 
-def register():
+
+def register() -> None:  # pragma: no cover - Blender registration
     for cls in classes:
         bpy.utils.register_class(cls)
 
-    # Only add properties if they don't already exist (safe for script reloads)
     if not hasattr(bpy.types.Scene, "dicomator_props"):
         bpy.types.Scene.dicomator_props = PointerProperty(type=DICOMatorProperties)
 
@@ -1306,8 +60,8 @@ def register():
             precision=0,
         )
 
-def unregister():
-    # Remove custom properties if present
+
+def unregister() -> None:  # pragma: no cover - Blender registration
     if hasattr(bpy.types.Scene, "dicomator_props"):
         del bpy.types.Scene.dicomator_props
 
@@ -1317,8 +71,12 @@ def unregister():
     for cls in reversed(classes):
         bpy.utils.unregister_class(cls)
 
-if __name__ == "__main__":
-    register()
-if __name__ == "__main__":
-    register()
 
+__all__ = [
+    "add_gaussian_noise",
+    "export_voxel_grid_to_dicom",
+    "voxelize_mesh",
+    "voxelize_objects_to_hu",
+    "register",
+    "unregister",
+]
